ai_code_review/requirements.txt
phidata>=2.0.0
openai>=1.3.0
pydantic>=2.5.0
python-dotenv>=1.0.0
pytest>=7.4.0
black>=23.11.0
pylint>=3.0.0
mypy>=1.7.0 